refactor: move mouse / toast / highlight into src/ui/
These three modules were sitting at `src/` root but are all UI-layer
concerns:

- `mouse.rs` — `ClickAction` + `HitTestRegistry`. Only rendering code
  registers click zones, only the event loop hit-tests them.
- `toast.rs` — cross-panel notification queue, drained by the status
  bar renderer. Never consumed outside `ui/`.
- `highlight.rs` — syntect-backed syntax coloring for file preview,
  only used by `ui/file_preview_panel.rs` and the `highlight` bench.

All three now live at `src/ui/`. `lib.rs` drops their top-level `pub
mod` declarations; `ui/mod.rs` adds them. Every `crate::mouse::` /
`crate::toast::` / `crate::highlight::` path rewritten to
`crate::ui::mouse::` / `crate::ui::toast::` / `crate::ui::highlight::`.
External callers (tests / benches) get `reef::ui::mouse::` etc.

Main.rs had a stale `use reef::ui::mouse;` that the batch rewrite
surfaced (only fully-qualified paths are used below) — removed.

124/124 tests, clippy + fmt clean.
