+++ src/app.rs
@@ -2,8 +2,8 @@ use crate::file_tree::{self, FileTree, PreviewContent};
 use crate::fs_watcher;
 use crate::git::graph::GraphRow;
 use crate::git::{CommitDetail, DiffContent, FileEntry, GitRepo, RefLabel};
-use crate::mouse::{ClickAction, HitTestRegistry};
-use crate::toast::Toast;
+use crate::ui::mouse::{ClickAction, HitTestRegistry};
+use crate::ui::toast::Toast;
 use std::collections::{HashMap, HashSet};
 use std::path::PathBuf;
 use std::sync::mpsc;
