+++ src/lib.rs
@@ -6,8 +6,5 @@ pub mod app;
 pub mod file_tree;
 pub mod fs_watcher;
 pub mod git;
-pub mod highlight;
-pub mod mouse;
 pub mod prefs;
-pub mod toast;
 pub mod ui;
